:recycle: rename back path parameter to 'uuid'
