+++ src/objects/api/fields.py
@@ -48,3 +48,15 @@ class ObjectTypeField(serializers.RelatedField):
 
     def to_representation(self, obj):
         return obj.url
+
+
+class ObjectUrlField(serializers.HyperlinkedIdentityField):
+    lookup_field = "uuid"
+
+    def get_url(self, obj, view_name, request, format):
+        if hasattr(obj, "pk") and obj.pk in (None, ""):
+            return None
+
+        lookup_value = getattr(obj.object, "uuid")
+        kwargs = {self.lookup_url_kwarg: lookup_value}
+        return self.reverse(view_name, kwargs=kwargs, request=request, format=format)
