+++ src/objects/api/v2/filters.py
@@ -1,5 +1,3 @@
-import datetime
-
 from django import forms
 from django.utils.translation import ugettext_lazy as _
