+++ src/objects/api/v2/views.py
@@ -3,14 +3,10 @@ import datetime
 from django.conf import settings
 from django.db import models
 
-from drf_spectacular.utils import (
-    OpenApiParameter,
-    OpenApiTypes,
-    extend_schema,
-    extend_schema_view,
-)
+from drf_spectacular.utils import extend_schema, extend_schema_view
 from rest_framework import mixins, viewsets
 from rest_framework.decorators import action
+from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 from vng_api_common.search import SearchMixin
 
@@ -65,6 +61,7 @@ class ObjectViewSet(
     serializer_class = ObjectSerializer
     filterset_class = ObjectRecordFilterSet
     lookup_field = "object__uuid"
+    lookup_url_kwarg = "uuid"
     search_input_serializer_class = ObjectSearchSerializer
     permission_classes = [ObjectTypeBasedPermission]
     pagination_class = DynamicPageSizePagination
@@ -112,7 +109,7 @@ class ObjectViewSet(
         responses={"200": HistoryRecordSerializer(many=True)},
     )
     @action(detail=True, methods=["get"], serializer_class=HistoryRecordSerializer)
-    def history(self, request, object__uuid=None):
+    def history(self, request, uuid=None):
         """Retrieve all RECORDs of an OBJECT."""
         records = self.get_object().object.records.order_by("id")
         serializer = self.get_serializer(records, many=True)
